fix cyclic dependencies enough for now
