+++ extension/src/background/activePilotSessions.ts
@@ -1,0 +1,4 @@
+import { PilotSession } from './types'
+
+/** maps `windowId` to pilot session */
+export const activePilotSessions = new Map<number, PilotSession>()
