+++ extension/src/components/ConnectionStack/index.tsx
@@ -1,15 +1,15 @@
 import cn from 'classnames'
 import React from 'react'
-
-import { MODULE_NAMES } from '../../panel/integrations/zodiac/useZodiacModules'
+import { ChainId } from 'ser-kit'
+import { MODULE_NAMES } from '../../const'
 import { LegacyConnection } from '../../types'
 import Address from '../Address'
 import Box from '../Box'
-
 import classes from './style.module.css'
 
 interface Props {
   connection: LegacyConnection
+  chainId: ChainId
   helperClass?: string
   addressBoxClass?: string
   className?: string
@@ -17,6 +17,7 @@ interface Props {
 
 const ConnectionStack: React.FC<Props> = ({
   connection,
+  chainId,
   helperClass,
   addressBoxClass,
   className,
@@ -25,7 +26,7 @@ const ConnectionStack: React.FC<Props> = ({
   return (
     <div className={cn(classes.connectionStack, className)}>
       <Box rounded className={cn([classes.address, addressBoxClass])}>
-        <Address address={pilotAddress} />
+        <Address chainId={chainId} address={pilotAddress} />
 
         <div className={cn(classes.helper, helperClass)}>
           <p>Pilot Account</p>
@@ -33,14 +34,14 @@ const ConnectionStack: React.FC<Props> = ({
       </Box>
       {moduleAddress && (
         <Box roundedRight className={cn([classes.address, addressBoxClass])}>
-          <Address address={moduleAddress} />
+          <Address chainId={chainId} address={moduleAddress} />
           <div className={cn(classes.helper, helperClass)}>
             <p>{(moduleType && MODULE_NAMES[moduleType]) || 'Zodiac'} Mod</p>
           </div>
         </Box>
       )}
       <Box roundedRight className={cn([classes.address, addressBoxClass])}>
-        <Address address={avatarAddress} />
+        <Address chainId={chainId} address={avatarAddress} />
         {avatarAddress && (
           <div className={cn(classes.helper, helperClass)}>
             <p>Piloted Safe</p>
