+++ extension/src/components/index.ts
@@ -11,7 +11,6 @@ export { default as ConnectionStack } from './ConnectionStack'
 export { default as Field } from './Field'
 export { default as Flex } from './Flex'
 export { default as IconButton } from './IconButton'
-export { RouteBubble } from './RouteBubble'
 export { Select, selectStyles } from './Select'
 export { default as Spinner } from './Spinner'
 export { default as Tag } from './Tag'
