+++ extension/src/const.ts
@@ -1,0 +1,7 @@
+import { KnownContracts } from '@gnosis.pm/zodiac'
+
+export const MODULE_NAMES = {
+  [KnownContracts.DELAY]: 'Delay',
+  [KnownContracts.ROLES_V1]: 'Roles v1',
+  [KnownContracts.ROLES_V2]: 'Roles v2',
+}
