+++ extension/src/panel/state/actions.ts
@@ -1,6 +1,6 @@
 import { MetaTransactionData } from '@safe-global/safe-core-sdk-types'
 import { ContractInfo } from '../utils/abi'
-import { ExecutionStatus } from './reducer'
+import { ExecutionStatus } from './executionStatus'
 
 interface AppendTransactionAction {
   type: 'APPEND_TRANSACTION'
