+++ extension/src/panel/state/index.tsx
@@ -28,3 +28,5 @@ export const ProvideState: React.FC<{ children: ReactNode }> = ({
     </DispatchContext.Provider>
   )
 }
+
+export { ExecutionStatus } from './executionStatus'
