+++ extension/src/panel/transactions/RouteBubble/ConnectionsIcon.tsx
